backend/requirements.txt
addict
anthropic
cryptography>=41.0.0  # For JWT and encryption support
elevenlabs>=2.1.0
flask
beautifulsoup4
# google-api-python-client
# google-auth-httplib2
# google-auth-oauthlib
# google-cloud-storage
gjdutils
gunicorn
jinja2
loguru>=0.7.2  # For simplified logging
openai
peewee>=3.17.0
peewee-migrate>=1.12.2
pendulum
Pillow
psutil>=5.9.0
pycountry
pydantic
PyJWT>=2.8.0  # For JWT token handling
python-slugify
psycopg2-binary==2.9.9
python-dotenv>=1.0.0  # For environment variable management
requests>=2.31.0  # For HTTP requests to Supabase
# spacy>=3.8.0  # For NLP processing - not actually used in the codebase
ujson>=5.10.0  # For JSON handling
yt-dlp>=2023.12.30  # For YouTube audio downloads
flask-cors>=4.0.0  # For cross-origin API access
whitenoise>=6.6.0  # For static file serving on Vercel
# pytest-postgresql>=5.0.0
PyICU>=2.12  # ICU-based multilingual segmentation
